Fix for multiple headers
