+++ libhttpparser/test/HttpRequestParserTest.cpp
@@ -66,10 +66,92 @@ struct HttpRequestTestData {
 					  requestChunks.begin(), requestChunks.end(), 0, [](int sum, const std::string& str) { return sum + str.length(); }))) {
 	}
 };
+using namespace ::testing;
 
-class HttpRequestParserTest : public ::testing::TestWithParam<HttpRequestTestData> {};
+class HttpRequestParserTest : public Test, public httpparser::HttpRequestParser {
+public:
+};
 
-TEST_P(HttpRequestParserTest, TestValidRequest) {
+TEST_F(HttpRequestParserTest, testParseXForwardedFor) {
+	std::string dataToTest = "fe80:0000:0000:0000:0000:0000:0000:0005";
+	parseClientIPValue(dataToTest);
+	EXPECT_EQ(result.clientIp.size(), 1);
+	EXPECT_EQ("fe80:0000:0000:0000:0000:0000:0000:0005", result.clientIp[0]);
+
+	result.clientIp.clear();
+
+	dataToTest = "203.0.113.195, 70.41.3.18, 150.172.238.178";
+	parseClientIPValue(dataToTest);
+	EXPECT_EQ(result.clientIp.size(), 3);
+	EXPECT_EQ("203.0.113.195", result.clientIp[0]);
+	EXPECT_EQ("70.41.3.18", result.clientIp[1]);
+	EXPECT_EQ("150.172.238.178", result.clientIp[2]);
+
+	result.clientIp.clear();
+
+	dataToTest = "203.0.113.195";
+	parseClientIPValue(dataToTest);
+	EXPECT_EQ(result.clientIp.size(), 1);
+	EXPECT_EQ("203.0.113.195", result.clientIp[0]);
+
+	result.clientIp.clear();
+
+	dataToTest = "2001:db8:85a3:8d3:1319:8a2e:370:7348";
+	parseClientIPValue(dataToTest);
+	EXPECT_EQ(result.clientIp.size(), 1);
+	EXPECT_EQ("2001:db8:85a3:8d3:1319:8a2e:370:7348", result.clientIp[0]);
+
+	result.clientIp.clear();
+
+	dataToTest = "203.0.113.195:41237, 198.51.100.100:38523";
+	parseClientIPValue(dataToTest);
+	EXPECT_EQ(result.clientIp.size(), 2);
+	EXPECT_EQ("203.0.113.195", result.clientIp[0]);
+	EXPECT_EQ("198.51.100.100", result.clientIp[1]);
+
+	result.clientIp.clear();
+
+	dataToTest = "[2001:db8::1a2b:3c4d]:41237, 198.51.100.100:26321";
+	parseClientIPValue(dataToTest);
+	EXPECT_EQ(result.clientIp.size(), 2);
+	EXPECT_EQ("2001:db8::1a2b:3c4d", result.clientIp[0]);
+	EXPECT_EQ("198.51.100.100", result.clientIp[1]);
+
+	result.clientIp.clear();
+
+	dataToTest = "[2001:db8::aa:bb]";
+	parseClientIPValue(dataToTest);
+	EXPECT_EQ(result.clientIp.size(), 1);
+	EXPECT_EQ("2001:db8::aa:bb", result.clientIp[0]);
+
+	result.clientIp.clear();
+
+	dataToTest = "203.0.113.195, 2001:db8:85a3:8d3:1319:8a2e:370:7348";
+	parseClientIPValue(dataToTest);
+	EXPECT_EQ(result.clientIp.size(), 2);
+	EXPECT_EQ("203.0.113.195", result.clientIp[0]);
+	EXPECT_EQ("2001:db8:85a3:8d3:1319:8a2e:370:7348", result.clientIp[1]);
+
+	result.clientIp.clear();
+
+	dataToTest = "203.0.113.195,2001:db8:85a3:8d3:1319:8a2e:370:7348,198.51.100.178";
+	parseClientIPValue(dataToTest);
+	EXPECT_EQ(result.clientIp.size(), 3);
+	EXPECT_EQ("203.0.113.195", result.clientIp[0]);
+	EXPECT_EQ("2001:db8:85a3:8d3:1319:8a2e:370:7348", result.clientIp[1]);
+	EXPECT_EQ("198.51.100.178", result.clientIp[2]);
+
+	result.clientIp.clear();
+
+	dataToTest = "[2001:db8::1]:30943";
+	parseClientIPValue(dataToTest);
+	EXPECT_EQ(result.clientIp.size(), 1);
+	EXPECT_EQ("2001:db8::1", result.clientIp[0]);
+}
+
+class HttpRequestDataParsingTest : public TestWithParam<HttpRequestTestData> {};
+
+TEST_P(HttpRequestDataParsingTest, TestValidRequest) {
 	const auto& testData{GetParam()};
 	HttpRequestParser parser;
 	size_t totalBytesParsed{0};
@@ -110,7 +192,7 @@ TEST_P(HttpRequestParserTestInvalid, testInvalidRequest) {
 
 INSTANTIATE_TEST_SUITE_P(
 		Default,
-		HttpRequestParserTest,
+		HttpRequestDataParsingTest,
 		::testing::Values(
 				HttpRequestTestData{
 						{"GET /example HTTP/1.1\r\nHost: example.com\r\n\r\n"}, "GET", "/example", "HTTP/1.1", "example.com", {}},
@@ -202,3 +284,4 @@ INSTANTIATE_TEST_SUITE_P(
 				HttpRequestTestInvalidData{{"\nGET / HTTP/1.1\r\nHost: example.com\r\n\r\n"}, 1},
 				HttpRequestTestInvalidData{{"GET / HTTP/0.0\r\nHost: example.com\r\n\r\n"}, 12},
 				HttpRequestTestInvalidData{{"GET http://example.com HTTP/1.1\r\nHost: example.com\r\n\r\n"}, 5}));
+
